==> Scripts/2_Reconcile.ipynb <==
# Implement reconciliation

import pandas as pd
import numpy as np
import time

start_time= time.time()
data = pd.read_csv('../Data/Trade_filtered_code.csv', sep=',', encoding='utf-8',index_col=0).reset_index()
data.columns = data.columns.str.lower().str.replace(' ', '_')
data

filter_t = data.item.unique()

# Define origin & destination
to_add = data.loc[:,['reporter_country_code','reporter_countries','partner_country_code','partner_countries']]
to_add.columns = ['origin_country_code','origin_country','destin_country_code','destin_country']
data=data.join(to_add)

replace_bool = (data.element == 'Import Quantity') | (data.element == 'Import Value')
data.loc[replace_bool, ['origin_country', 'origin_country_code']] = data.loc[replace_bool, ['partner_countries', 'partner_country_code']].values
data.loc[replace_bool, ['destin_country', 'destin_country_code']] = data.loc[replace_bool, ['reporter_countries', 'reporter_country_code']].values

data.head(10)

### Reconciliation
Find repeated instances of code.

# takes a bit good (3min)
data['double_reports']= 1*(data.groupby(["origin_country",'destin_country',"year","item","unit"], group_keys=False)['year'].transform('count')>1)
data['same_double_reports']= 1*(data.groupby(["origin_country",'destin_country',"year",'item','unit','value'], group_keys=False)["value"].transform('count')>1)

# Reliability estimation: (1.2 min)
reliability=data.groupby("reporter_countries", group_keys=False).apply(lambda x: sum(x.loc[:,'same_double_reports']) /sum(x.loc[:,'double_reports']))
reliability=pd.DataFrame(reliability,columns=['reliability'])

data=pd.merge(data,reliability, left_on='reporter_countries',right_index=True)
print('Cum time (Reliability):', (time.time()- start_time)/60,'min')

# Remove where there is a double report and the country is not the preferred one 
# Find reliable country (can take up to 5min)

reliable_clean = data.sort_index().groupby(["origin_country",'destin_country',"year",'item','unit'], group_keys=False).tail(1)
reliable_clean=reliable_clean.drop(columns=['reporter_country_code', 'reporter_countries','partner_country_code','partner_countries','reliability','double_reports','same_double_reports'])

# Drop Paddy rice column
conditon= reliable_clean.item == 'Rice, paddy (rice milled equivalent)'
reliable_clean = reliable_clean.loc[~conditon,:]

reliable_clean.to_csv('../Data/Trade_reconciled_code.csv',encoding='utf-8',index=False) # Checked data clean same dimensions than R code need to check values
print('Cum time (after save):', (time.time()- start_time)/60,'min')
reliable_clean.head(10)  

# Compare to R dataset
reliable_clean = pd.read_csv('../Data/Trade_reconciled_code.csv', sep=',', encoding='utf-8',index_col='index').reset_index()
reliable_clean['item']=(reliable_clean.item.str.lower())
#                        .replace('matã©','mate',regex=True)
#                        .replace(';',',',regex=True)
#                        .replace('"','',regex=True).str.lower())# Separator is a coma

r_dataset = pd.read_csv('../FAO_2023_10.csv', sep=',', encoding='utf-8').reset_index()
r_dataset['item']=(r_dataset.item.str.lower())
#                        .replace('maté','mate',regex=True)
#                        .replace(';',',',regex=True)
#                        .replace('"','',regex=True).str.lower())# Separator is a coma

# Validate 
print('Items R:',len(r_dataset.item.unique()))
print('Items python:',len(reliable_clean.item.unique()))

#Differences in items shown
set(reliable_clean.item.unique()).difference(set(r_dataset.item.unique()))

print('Countries R:',len(set(r_dataset.origin_country_code.unique()).union(r_dataset.destin_country_code.unique())))

print('Countries python:',len(set(reliable_clean.origin_country_code.unique()).union(reliable_clean.destin_country_code.unique())))

print('Samples R:',len(r_dataset))
print('Samples python:',len(reliable_clean))
