==> Scripts/6_Explore_production.ipynb <==
# Explore production and stock data
import pandas as pd
import numpy as np
import time
import matplotlib.pyplot as plt
import pickle

# Production data Load  
data_prod= pd.read_csv('../Data/Production_Crops_Livestock_E_All_Data_(Normalized)/Production_Crops_Livestock_E_All_Data_(Normalized).csv', sep=',',low_memory=False, encoding='latin1',index_col=0).reset_index()
data_prod.columns = data_prod.columns.str.lower().str.replace(' ', '_')

flag_proc = 'processed'

if flag_proc == 'processed':
    data_trade = pd.read_pickle('../Data/Data_food_groups_code.pkl').reset_index()
    all_countries = set(data_trade.origin_country).union(data_trade.destin_country)
    print('Using processed dataset (Data_food_groups_code.pkl)')
else: 
    data_trade = pd.read_csv('../Data/raw_trade/Trade_DetailedTradeMatrix_E_All_Data_(Normalized).csv', sep=',', encoding='latin-1',index_col=0).reset_index()
    data_trade.columns = data_trade.columns.str.lower().str.replace(' ', '_')
    all_countries = set(data_trade.reporter_country).union(data_trade.partner_country)
    print('Using raw dataset (Trade_DetailedTradeMatrix_E_All_Data_(Normalized).csv)')


data_trade

# FILTER PRODUCTION: remove areas that are not countries
data_prod_countries = data_prod.loc[(data_prod.area.isin(all_countries)),:]
len(data_prod.area.unique())

drop_geom = set(data_prod.area.unique()).difference(all_countries)


def Standardise_names(data):
    # Fix formatting isseues 
    data.loc[:,'area']= data.loc[:,'area'].replace('Micronesia','Micronesia (Federated States of)')
    data.loc[:,'area']=data.loc[:,'area'].replace('China','China, mainland')
    data.loc[:,'area']=data.loc[:,'area'].replace('Polynesia','French Polynesia')
    data.loc[:,'area']=data.loc[:,'area'].replace('Palestine','State of Palestine')
    data.loc[:,'area']=data.loc[:,'area'].replace("Ethiopia PDR",'Ethiopia')
    
    return data

# Standardise country names
data_prod_std = Standardise_names(data_prod_countries)

# Keep only years in TRADE data 
data_prod_std = data_prod_std.loc[(data_prod_std.year>=1986) & (data_prod_std.year<=2021),:]

print('Num of "item_codes" in data_prod:',len(data_prod_std.item_code.unique()))
print('Num of "item" in data_prod:',len(data_prod_std.item.unique()))

print('Num of "item_code" in data_trade:',len(data_trade.item_code.unique()))
print('Num of "item" in data_trade,', len(data_trade.item.unique()))

print('Food products in data_prod AND in data_trade:',len(set(data_prod_std.item_code.unique()).intersection(data_trade.item_code.unique())))
print('Food products in data_prod AND in data_trade:',len(set(data_prod_std.item.unique()).intersection(data_trade.item.unique())))

not_intersected = set(data_trade.item).difference(data_prod_std.item.unique())
print('Food products in data_trade not in data_prod:',len(not_intersected))

not_intersected

not_intersected = set(data_prod_std.item).difference(data_trade.item.unique())
#not_intersected = set(data_prod.item_code).difference(data_trade.item_code.unique())

print('Food products in data_prod not in data_trade:',len(not_intersected))
not_intersected

# KEEP BASIC PRODUCTS: 
to_keep = set(data_prod_std.item.unique()).intersection(data_trade.item.unique()) # keep products in 

data_prod_keep = data_prod_std.loc[data_prod_std.item.isin(to_keep),:]
data_trade_keep = data_trade.loc[(data_trade.item.isin(to_keep)),:] #& (data_trade.unit =='1000 US$')

# Keep only production column
data_prod_keep = data_prod_keep.loc[(data_prod_keep.element=='Production'),:].reset_index(drop=True)
data_prod_keep= data_prod_keep.drop(columns =['area_code_(m49)','item_code_(cpc)','element_code','year_code','flag','note'])

# Check temporal resolution of data: 
group_y= data_prod_keep.groupby(by='item_code').apply(lambda group : len(group.year.unique()))
print(group_y.value_counts())

# Drop low temporal resolution products (for tonnes): "Rubber..." and "Butter goat milk"
data_prod_keep= data_prod_keep[~data_prod_keep.item_code.isin([839,1022])] 
data_trade_keep= data_trade_keep[data_trade_keep.item_code.isin(data_prod_keep.item_code.unique())]  
# 216 products that have complete temporal ressolution beteween 1986-2021

data_trade_keep

len(data_trade_keep.item.unique())

pickle.dump(data_trade_keep, open('../Data/Data_trade_filt_production.pkl', 'wb'))


## Explore most traded products

# Do top for 2019
data_trade = pd.read_pickle('../Data/Data_trade_filt_production.pkl').reset_index()  #drop =True

data_trade_2019= data_trade.loc[data_trade.year==2019,:]
data_item_sum = data_trade.groupby(['item','unit'])['value'].sum()

# Pareto rule: 80% of the trade on the hands of 20%. 90%trade = 46 products, 80%trade = 21 products

sort_products = data_item_sum.xs('tonnes', level='unit').sort_values(ascending =False)
sort_products

#
cum_fraction = 0.8
relative_cumsum = (sort_products/sort_products.sum()).cumsum()

sort_products= pd.DataFrame(sort_products)
sort_products['in_thr'] = 'blue'
sort_products.loc[relative_cumsum<=cum_fraction,'in_thr']='orange'
# Keep subset
subset_products_all = sort_products[relative_cumsum<=cum_fraction]
#subset_products_tonnes = sort_products[relative_cumsum<=cum_fraction]

#pickle.dump(subset_products, open('../Data/Data_trade_filt_production_top.pkl', 'wb'))


# 80% tonnes : 21 products. 
# 80% dollars: 45 products
# 20/21 products in tonnes are also in the 80% for dollars, the only one that is not there is "Molasses".

set(subset_products.index).difference(subset_products_all.index)

print('Subset of products with ',cum_fraction*100,'% of trade (tonnes): ',len(subset_products))
print('Fraction of all products: ',100*len(subset_products)/len(sort_products),'%')

sort_products['cumsum_val'] = sort_products.value.cumsum()
sort_products


# 
fig, ax = plt.subplots (figsize=(10,5))
ax.bar(sort_products.index,sort_products.cumsum_val,color =sort_products.in_thr)
ax.tick_params(axis='x', labelrotation=90)
ax.tick_params(right= True,top= True,left= True, bottom= True)

ax.set_title('Food items total exports (Tonnes)')
#ax.axhline(1e6,color='dimgray')
ax.set_xticklabels([])

fig, ax = plt.subplots (figsize=(10,5))
ax.bar(sort_products.index,sort_products.value,color =sort_products.in_thr)
ax.tick_params(axis='x', labelrotation=90)
ax.tick_params(right= True,top= True,left= True, bottom= True)

ax.set_title('Food items total exports (Tonnes)')
#ax.axhline(1e6,color='dimgray')
ax.set_xticklabels([])

plt.show()

fig, ax = plt.subplots (1)
ax.hist(sort_products,bins =500)


trheshold = 1e6
sort_products = pd.DataFrame(data_item_sum.xs('tonnes', level='unit').sort_values(ascending =False))
sort_products['above_thr'] = 'blue'
sort_products.loc[sort_products.value>= trheshold,'above_thr'] = 'orange'

fig, ax = plt.subplots (figsize=(10,5))
ax.bar(sort_products.index,sort_products.value,color =sort_products.above_thr)
ax.tick_params(axis='x', labelrotation=90)
ax.tick_params(right= True,top= True,left= True, bottom= True)

ax.set_title('Food items total exports (Tonnes)')
#ax.axhline(1e6,color='dimgray')
ax.set_xticklabels([])

plt.show()

#Check top 50 products:



len(sort_products[sort_products<1e6])

more1m = sort_products[sort_products>=1e6]

len(more1m)
